Parse datasec

+++ types/mut.hpp
@@ -3,7 +3,7 @@
 
 #include <cstdint>
 
-enum Mut : int32_t { const_ = 0x00, var_ = 0x01, invalid_mut = 0xAA };
+enum Mut : uint8_t { const_ = 0x00, var_ = 0x01, invalid_mut = 0xAA };
 
 namespace antiwasm {
 /**
