+++ types/reftype.hpp
@@ -8,7 +8,7 @@ namespace antiwasm {
 
 const auto REFTYPE_SIZE = 1;
 
-enum Reftype : uint32_t { funref = 0x70, externref = 0x6F, invalid_reftype = 0xFF };
+enum Reftype : uint8_t { funref = 0x70, externref = 0x6F, invalid_reftype = 0xFF };
 
 /**
  * Parses and returns the reftype value.
